requirements.txt
# Feedparser 5.1.3 raises an encoding exception on Python 3
# See https://code.google.com/p/feedparser/issues/detail?id=384
-e git+https://code.google.com/p/feedparser/#egg=feedparser
PyYAML

# For testing.
pytest
flake8
nose
coveralls
